requirements.txt
asgiref==3.8.1
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.0
django-tenants==3.7.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0
drf-yasg>=1.21.7
inflection==0.5.1
packaging==25.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT>=2.0,<3.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
